put new test_data.py code in proper place
